[ts] Refactor Node impl to use smaller packages
Start refactoring the project into smaller packages. Move type definitions and some tests to shared packages.
