+++ package.json
@@ -35,6 +35,7 @@
   },
   "workspaces": {
     "packages": [
+      "packages/*",
       "node",
       "ts"
     ]
