+++ packages/core/.gitignore
@@ -1,0 +1,2 @@
+lib/
+test/
