+++ packages/core/README.md
@@ -1,0 +1,1 @@
+# Core type definitions for `v8-coverage-tools`
