+++ packages/core/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "@v8-coverage-tools/core",
+  "private": true,
+  "version": "0.0.0",
+  "description": "Test specifications for Node.",
+  "author": "Charles Samborski <demurgos@demurgos.net> (https://demurgos.net)",
+  "license": "MIT",
+  "type": "module",
+  "files": [
+    "./lib/",
+    "./src/lib/"
+  ],
+  "exports": {
+    ".": "./lib/index.js",
+    "./*": "./lib/*.js"
+  },
+  "types": "./lib/index.d.ts",
+  "scripts": {
+    "build": "tsc --build src/lib/tsconfig.json"
+  },
+  "engines": {
+    "node": "^12.20.0 || >=14.13.1"
+  },
+  "devDependencies": {
+    "@types/chai": "^4.2.22",
+    "@types/mocha": "^9.0.0",
+    "@types/node": "^16.10.1",
+    "chai": "^4.3.4",
+    "mocha": "^9.1.2",
+    "typescript": "^4.4.3"
+  }
+}
