+++ packages/mocha/README.md
@@ -1,0 +1,1 @@
+# Mocha test suite generator for v8-coverage-tools
