+++ ts/package.json
@@ -39,10 +39,14 @@
   "engines": {
     "node": "^12.20.0 || >=14.13.1"
   },
+  "dependencies": {
+    "@v8-coverage-tools/core": "workspace:*"
+  },
   "devDependencies": {
     "@types/chai": "^4.2.22",
     "@types/mocha": "^9.0.0",
     "@types/node": "^16.10.1",
+    "@v8-coverage-tools/mocha": "workspace:*",
     "chai": "^4.3.4",
     "mocha": "^9.1.2",
     "typescript": "^4.4.3"
